+++ bundles/mapping/mapwfs2/plugin/components/FeatureUtil.js
