+++ bundles/mapping/mapwfs2/plugin/components/defaultStyle.js
